Clean up main.py
